mathongo-ai-qc/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
langchain==0.2.5
langchain-core==0.2.10
langgraph==0.0.68
google-generativeai==0.7.0
langchain-google-genai==1.0.6
python-dotenv==1.0.1
pydantic==2.7.4
